refactor: modularize logger plugin implementation (724→<300 LOC per file)
Split large logger module into focused, single-responsibility components:
- log_severity.rs (84 LOC) - log severity enum with comprehensive tests
- log_status.rs (133 LOC) - LogStatus structure for osquery status logs
- logger_features.rs (56 LOC) - logger feature flags and constants
- logger_plugin.rs (189 LOC) - LoggerPlugin trait with default implementations
- logger_wrapper.rs (361 LOC) - LoggerPluginWrapper for osquery integration
- mod.rs (65 LOC) - module organization and documentation

Each component follows SRP with 27 comprehensive unit tests. All osquery
logger protocol functionality preserved while improving code organization.
